+++ src/object/selectValues.js
@@ -25,7 +25,7 @@ import { getMatcher } from '../util/matchers';
  *
  *   Object.selectValues({a:1,b:2}, 1) -> {a:1}
  *   Object.selectValues({a:1,b:1}, 3) -> {}
- *   Object.selectValues({a:'a',b:'b'}, /[a-f]) -> {a:'a',b:'b'}
+ *   Object.selectValues({a:'a',b:'b'}, /[a-f]/) -> {a:'a',b:'b'}
  *   Object.selectValues(usersById, user => {
  *     return user.age > 30;
  *   }); -> an object with all users over 30